[🪽 FEAT] 게임 전적 연결 , 게임 결과 창 리디자인 및 백엔드 로직 연결, 마이페이지 주기적 데이터 요청 로직 추가
