+++ handtris/src/util/getRoomCode.ts
@@ -1,3 +1,7 @@
 export const getRoomCode = (): string | null => {
-  return sessionStorage.getItem("roomCode") || localStorage.getItem("roomCode");
+  return sessionStorage.getItem("roomCode") || null;
+};
+
+export const getRoomName = (): string | null => {
+  return sessionStorage.getItem("roomName") || null;
 };
